--- rmb_tests/collections/ansible_collections/myorg/myos/plugins/module_utils/network/myos/config/interfaces/interfaces.py
@@ -1,96 +1,0 @@
-#
-# -*- coding: utf-8 -*-
-# Copyright 2020 Red Hat
-# GNU General Public License v3.0+
-# (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
-#
-
-from __future__ import absolute_import, division, print_function
-__metaclass__ = type
-
-"""
-The myos_interfaces config file.
-It is in this file where the current configuration (as dict)
-is compared to the provided configuration (as dict) and the command set
-necessary to bring the current configuration to its desired end-state is
-created.
-"""
-
-from copy import deepcopy
-
-from ansible.module_utils.six import iteritems
-from ansible_collections.ansible.netcommon.plugins.module_utils.network.common.utils import (
-    dict_merge,
-)
-from ansible_collections.ansible.netcommon.plugins.module_utils.network.common.resource_module import (
-    ResourceModule,
-)
-from ansible_collections.myorg.myos.plugins.module_utils.network.myos.nxos.facts.facts import (
-    Facts,
-)
-from ansible_collections.myorg.myos.plugins.module_utils.network.myos.rm_templates.interfaces import (
-    InterfacesTemplate,
-)
-
-
-class Interfaces(ResourceModule):
-    """
-    The myos_interfaces config class
-    """
-
-    def __init__(self, module):
-        super(Interfaces, self).__init__(
-            empty_fact_val={},
-            facts_module=Facts(module),
-            module=module,
-            resource="interfaces",
-            tmplt=InterfacesTemplate(),
-        )
-        self.parsers = [
-        ]
-
-    def execute_module(self):
-        """ Execute the module
-
-        :rtype: A dictionary
-        :returns: The result from module execution
-        """
-        if self.state not in ["parsed", "gathered"]:
-            self.generate_commands()
-            self.run_commands()
-        return self.result
-
-    def generate_commands(self):
-        """ Generate configuration commands to send based on
-            want, have and desired state.
-        """
-        wantd = {}
-        haved = {}
-
-        # if state is merged, merge want onto have and then compare
-        if self.state == "merged":
-            wantd = dict_merge(haved, wantd)
-
-        # if state is deleted, empty out wantd and set haved to wantd
-        if self.state == "deleted":
-            haved = {
-                k: v for k, v in iteritems(haved) if k in wantd or not wantd
-            }
-            wantd = {}
-
-        # remove superfluous config for overridden and deleted
-        if self.state in ["overridden", "deleted"]:
-            for k, have in iteritems(haved):
-                if k not in wantd:
-                    self._compare(want={}, have=have)
-
-        for k, want in iteritems(wantd):
-            self._compare(want=want, have=haved.pop(k, {}))
-
-    def _compare(self, want, have):
-        """Leverages the base class `compare()` method and
-           populates the list of commands to be run by comparing
-           the `want` and `have` data with the `parsers` defined
-           for the Interfaces network resource.
-        """
-        self.compare(parsers=self.parsers, want=want, have=have)
