--- rmb_tests/roles/tests/tasks/main.yml
@@ -1,4 +1,0 @@
-- include: "{{ item }}"
-  loop:
-  - facts.yml
-  - argspec.yml
