--- rmb_tests/test.yml
@@ -1,11 +1,0 @@
-- hosts: myos101
-  module_defaults:
-    myorg.myos.myos_facts:
-      gather_subset: min
-      gather_network_resources:
-      - interfaces
-  gather_facts: True
-  collections:
-  - myorg.myos
-  roles:
-  - tests
