+++ tests/rmb_tests/collections/ansible_collections/myorg/myos/LICENSE
