+++ tests/rmb_tests/collections/ansible_collections/myorg/myos/plugins/module_utils/network/myos/facts/interfaces/interfaces.py
@@ -1,0 +1,84 @@
+# -*- coding: utf-8 -*-
+# Copyright 2020 Red Hat
+# GNU General Public License v3.0+
+# (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+from __future__ import absolute_import, division, print_function
+__metaclass__ = type
+
+"""
+The myos interfaces fact class
+It is in this file the configuration is collected from the device
+for a given resource, parsed, and the facts tree is populated
+based on the configuration.
+"""
+import q
+from copy import deepcopy
+
+from ansible.module_utils.six import iteritems
+from ansible_collections.ansible.netcommon.plugins.module_utils.network.common import (
+    utils,
+)
+from ansible_collections.myorg.myos.plugins.module_utils.network.myos.rm_templates.interfaces import (
+    InterfacesTemplate,
+)
+from ansible_collections.myorg.myos.plugins.module_utils.network.myos.argspec.interfaces.interfaces import (
+    InterfacesArgs,
+)
+
+class InterfacesFacts(object):
+    """ The myos interfaces facts class
+    """
+
+    def __init__(self, module, subspec='config', options='options'):
+        self._module = module
+        self.argument_spec = InterfacesArgs.argument_spec
+        spec = deepcopy(self.argument_spec)
+        if subspec:
+            if options:
+                facts_argument_spec = spec[subspec][options]
+            else:
+                facts_argument_spec = spec[subspec]
+        else:
+            facts_argument_spec = spec
+
+        self.generated_spec = utils.generate_dict(facts_argument_spec)
+
+    def populate_facts(self, connection, ansible_facts, data=None):
+        """ Populate the facts for Interfaces network resource
+
+        :param connection: the device connection
+        :param ansible_facts: Facts dictionary
+        :param data: previously collected conf
+
+        :rtype: dictionary
+        :returns: facts
+        """
+        facts = {}
+        objs = []
+
+        if not data:
+            # typically data is populated from the current device configuration
+            # data = connection.get('show running-config | section ^interface')
+            # using mock data instead
+            data = ("resource rsrc_a\n"
+                    "  a_bool\n"
+                    "  a_string choice_b\n"
+                    "resource rscrc_b\n"
+                    "  key is property01 value is value end\n"
+                    "  an_int 10\n")
+
+        # parse native config using the Interfaces template
+        interfaces_parser = InterfacesTemplate(lines=data.splitlines())
+        objs = list(interfaces_parser.parse().values())
+
+        ansible_facts['ansible_network_resources'].pop('interfaces', None)
+
+        params = utils.remove_empties(
+            utils.validate_config(self.argument_spec, {"config": objs})
+        )
+
+        facts['interfaces'] = params['config']
+        ansible_facts['ansible_network_resources'].update(facts)
+
+        return ansible_facts
