+++ tests/rmb_tests/inventory.yml
