+++ tests/rmb_tests/test.yml
@@ -1,0 +1,6 @@
+- hosts: localhost
+  gather_facts: True
+  collections:
+  - myorg.myos
+  roles:
+  - tests
